Rename TransactionContext to TransactionStatusContext
