+++ packages/ui/test/common/components/TransactionButton.test.tsx
@@ -9,7 +9,7 @@ import { TransactionButton } from '@/common/components/buttons/TransactionButton
 import { useSignAndSendTransaction } from '@/common/hooks/useSignAndSendTransaction'
 import { transactionMachine } from '@/common/model/machines'
 import { ApiContext } from '@/common/providers/api/context'
-import { TransactionContextProvider } from '@/common/providers/transaction/provider'
+import { TransactionStatusProvider } from '@/common/providers/transactionStatus/provider'
 
 import { getButton } from '../../_helpers/getButton'
 import { alice } from '../../_mocks/keyring'
@@ -92,12 +92,12 @@ describe('UI: TransactionButton', () => {
         <MockApolloProvider>
           <AccountsContext.Provider value={useAccounts}>
             <ApiContext.Provider value={api}>
-              <TransactionContextProvider>
+              <TransactionStatusProvider>
                 <TestButton />
                 <TransactionButton style="primary" size="large">
                   Start new transaction
                 </TransactionButton>
-              </TransactionContextProvider>
+              </TransactionStatusProvider>
             </ApiContext.Provider>
           </AccountsContext.Provider>
         </MockApolloProvider>
